backend/services/glucose-monitor-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
beautifulsoup4==4.12.2
django-redis==5.4.0